David Brownell <david-b@pacbell.net>:
Start updating the NOR flash coverage to use @deffn syntax, so the
commands have more consistent presentation and formatting.  This
reorganizes information and updates its presentation, except where
the information didn't really match the code.

This patch updates most of the driver specific support, creating one
new (and alphabetized!) section just for driver-specific data, where
previously that data was split over up to three sections.  Of note:

 - The at91sam7 docs were a bit out of date with respect to the code.

 - The "str9xpec" stuff still deserves some work.  For now, it sits
   in its own subsection; pretty messy.

 - Likewise the "mflash" stuff.  That's a parallel infrastructure,
   and is now in a section of its own.

 - The "mass_erase" commands for the Cortex M3 chips got turned into
   footnotes.  IMO, they should vanish sometime; they're superfluous.

 - There are still a bunch of undocumented NOR drivers.  Examples:
   avr(8), tms470, pic32mx, more.

Plus there are a handful of minor tweaks to the NAND docs (to help make
the NOR and NAND presentations be parallel); the "Command Index" has
been renamed as the "Command and Driver Index"; reference TI instead
of Luminary Micro in several places.


git-svn-id: svn://svn.berlios.de/openocd/trunk@1937 b42882b7-edfa-0310-969c-e2dbd0fdcd60